python

+++ python/购物问题.py
@@ -1,5 +1,5 @@
 thing = input('请输入你要购买的物品：')
-price = int (input("请输入你要购买的物品的价格："))
+price = float (input("请输入你要购买的物品的价格："))
 quan = int (input('请输入你要购买的物品的数量：'))
 area = price * quan
